Added marking of current item for aliases
This will fix the problem with marking aliases to current page. Before this, if there are two links to one menu item in menu (on to component and one alias to that item) only one is marked as current. This is wrong, both of them point to the same position so both should be marked as current.